# cmd/spec/openapi.yaml
components:
  schemas:
...
    v1.BadRequest:
      properties:
        Code:
          type: string
        Status:
          type: integer
        Title:
          type: string
      type: object
...
    v1.Image:
      properties:
        Account:
          type: string
        Commit:
          properties:
            Account:
              type: string
            Arch:
              type: string
            BlueprintToml:
              type: string
            BuildDate:
              type: string
            BuildNumber: {}
            ComposeJobID:
              type: string
            CreatedAt:
              format: date-time
              type: string
            DeletedAt: {}
            ID: {}
            ImageBuildHash:
              type: string
            ImageBuildParentHash:
              type: string
            ImageBuildTarURL:
              type: string
            InstalledPackages:
              items:
                properties:
                  CreatedAt:
                    format: date-time
                    type: string
                  DeletedAt: {}
                  ID: {}
                  UpdatedAt:
                    format: date-time
                    type: string
                  arch:
                    type: string
                  epoch:
                    type: string
                  name:
                    type: string
                  release:
                    type: string
                  sigmd5:
                    type: string
                  signature:
                    type: string
                  type:
                    type: string
                  version:
                    type: string
                type: object
              type: array
            OSTreeCommit:
              type: string
            OSTreeParentCommit:
              type: string
            OSTreeRef:
              type: string
            Packages:
              items:
                properties:
                  CreatedAt:
                    format: date-time
                    type: string
                  DeletedAt: {}
                  ID: {}
                  Name:
                    type: string
                  UpdatedAt:
                    format: date-time
                    type: string
                type: object
              type: array
            Status:
              type: string
            UpdatedAt:
              format: date-time
              type: string
          type: object
        CommitID: {}
        CreatedAt:
          format: date-time
          type: string
        DeletedAt: {}
        Description:
          type: string
        Distribution:
          type: string
        ID: {}
        ImageSetID: {}
        ImageType:
          type: string
        Installer:
          properties:
            Account:
              type: string
            Checksum:
              type: string
            ComposeJobID:
              type: string
            CreatedAt:
              format: date-time
              type: string
            DeletedAt: {}
            ID: {}
            ImageBuildISOURL:
              type: string
            SshKey:
              type: string
            Status:
              type: string
            UpdatedAt:
              format: date-time
              type: string
            Username:
              type: string
          type: object
        InstallerID: {}
        Name:
          type: string
        OutputTypes:
          items:
            type: string
          type: array
        Status:
          type: string
        UpdatedAt:
          format: date-time
          type: string
        Version:
          type: integer
      type: object
...
    v1.InternalServerError:
      properties:
        Code:
          type: string
        Status:
          type: integer
        Title:
          type: string
      type: object
...
    v1.Repo:
      properties:
        Commit:
          properties:
            Account:
              type: string
            Arch:
              type: string
            BlueprintToml:
              type: string
            BuildDate:
              type: string
            BuildNumber: {}
            ComposeJobID:
              type: string
            CreatedAt:
              format: date-time
              type: string
            DeletedAt: {}
            ID: {}
            ImageBuildHash:
              type: string
            ImageBuildParentHash:
              type: string
            ImageBuildTarURL:
              type: string
            InstalledPackages:
              items:
                properties:
                  CreatedAt:
                    format: date-time
                    type: string
                  DeletedAt: {}
                  ID: {}
                  UpdatedAt:
                    format: date-time
                    type: string
                  arch:
                    type: string
                  epoch:
                    type: string
                  name:
                    type: string
                  release:
                    type: string
                  sigmd5:
                    type: string
                  signature:
                    type: string
                  type:
                    type: string
                  version:
                    type: string
                type: object
              type: array
            OSTreeCommit:
              type: string
            OSTreeParentCommit:
              type: string
            OSTreeRef:
              type: string
            Packages:
              items:
                properties:
                  CreatedAt:
                    format: date-time
                    type: string
                  DeletedAt: {}
                  ID: {}
                  Name:
                    type: string
                  UpdatedAt:
                    format: date-time
                    type: string
                type: object
              type: array
            Status:
              type: string
            UpdatedAt:
              format: date-time
              type: string
          type: object
        CommitID: {}
        CreatedAt:
          format: date-time
          type: string
        DeletedAt: {}
        ID: {}
        RepoStatus:
          type: string
        RepoURL:
          type: string
        UpdatedAt:
          format: date-time
          type: string
      type: object
...
    v1.bool:
      type: boolean
info:
  license:
    name: MIT
  title: edge-api
  version: 1.0.0
openapi: 3.0.0
paths:
  /images:
    post:
      operationId: createImage
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/v1.Image'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/v1.Image'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/v1.BadRequest'
          description: The request sent couldn't be processed.
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/v1.InternalServerError'
          description: There was an internal server error.
      summary: Composes an image on Image Builder
    get:
      operationId: listImages
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  count:
                    type: integer
                    example: 100
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/v1.Image'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/v1.BadRequest'
          description: The request sent couldn't be processed.
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/v1.InternalServerError'
          description: There was an internal server error.
      summary: Lists all images for an account.
  /images/{imageId}/update:
    post:
      operationId: CreateImageUpdate
      parameters:
        - name: imageId
          in: path
          required: true
          description: ImageID
          schema:
            type : integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/v1.Image'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/v1.Image'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/v1.BadRequest'
          description: The request sent couldn't be processed.
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/v1.InternalServerError'
          description: There was an internal server error.
      summary: Composes an Update for a image
  /images/{imageId}:
    get:
      operationId: getImage
      parameters:
        - name: imageId
          in: path
          required: true
          description: ImageID
          schema:
            type : integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/v1.Image'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/v1.BadRequest'
          description: The request sent couldn't be processed.
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/v1.InternalServerError'
          description: There was an internal server error.
      summary: Get an image by id.
  /images/{ostreeCommitHash}/info:
    get:
      operationId: GetImageByOstree
      parameters:
        - name: ostreeCommitHash
          in: path
          required: true
          description: ostreeCommitHash
          schema:
            type : string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/v1.Image'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/v1.BadRequest'
          description: The request sent couldn't be processed.
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/v1.InternalServerError'
          description: There was an internal server error.
      summary: Get an image by ostree hash.    
  /images/{imageId}/status:
    get:
      operationId: getImageStatus
      parameters:
        - name: imageId
          in: path
          required: true
          description: ImageID
          schema:
            type : integer
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  Status:
                    type: string
                    example: BUILDING
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/v1.BadRequest'
          description: The request sent couldn't be processed.
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/v1.InternalServerError'
          description: There was an internal server error.
      summary: Get image status.
      description: This method goes to image builder if the image is still building and updates the status if needed.
  /images/{imageId}/repo:
    get:
      operationId: getImageRepo
      parameters:
        - name: imageId
          in: path
          required: true
          description: ImageID
          schema:
            type : integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/v1.Repo'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/v1.BadRequest'
          description: The request sent couldn't be processed.
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/v1.InternalServerError'
          description: There was an internal server error.
      summary: Get Image OSTree repo.
      description: Returns the information of the OSTree Repository of a particular Image.
...
  /updates/device/{DeviceUUID}/image:
    get:
      operationId: GetImageInfo
      parameters:
        - name: DeviceUUID
          in: path
          required: true
          description: DeviceUUID
          schema:
            type : string
      responses:
        "200":
          content:
            application/json:
              schema:
                type: object
                properties:
                  Image:
                    $ref: '#/components/schemas/v1.Image'
                  
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/v1.BadRequest'
          description: The request sent couldn't be processed.
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/v1.InternalServerError'
          description: There was an internal server error.
      summary: Return image running on device.
      description: Return image running on device.   
...
  /checkImageName:
    post:
      operationId: checkImageName
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/v1.Image'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/v1.bool'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/v1.BadRequest'
          description: The request sent couldn't be processed.
        "500":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/v1.InternalServerError'
          description: There was an internal server error.
      summary: Check if image name is in use for current account 